+++ packages/editor/adf-utils/src/__tests__/unit/validator/validate.ts
@@ -4,7 +4,7 @@ import { validator } from '../../../validator';
 const validate = validator();
 
 declare var __dirname: string;
-const BASE_DIR = `${__dirname}/../../../../../editor-common/src/__tests__/unit/json-schema/v1-reference`;
+const BASE_DIR = `${__dirname}/../../../../../adf-schema/src/__tests__/unit/json-schema/v1-reference`;
 
 const readFilesSync = (path: string) =>
   fs.readdirSync(path).reduce(