+++ packages/editor/adf-utils/src/builders/nodes/media-single.ts
@@ -2,7 +2,7 @@ import {
   MediaSingleDefinition,
   MediaDefinition,
   MediaSingleAttributes,
-} from '@atlaskit/editor-common';
+} from '@atlaskit/adf-schema';
 
 export const mediaSingle = (attrs: MediaSingleAttributes | undefined) => (
   content: MediaDefinition,