+++ packages/editor/editor-core/src/__tests__/unit/plugins/table/layout.ts
@@ -14,7 +14,7 @@ import {
   layoutSection,
   layoutColumn,
 } from '@atlaskit/editor-test-helpers';
-import { TableLayout } from '@atlaskit/editor-common';
+import { TableLayout } from '@atlaskit/adf-schema';
 import {
   pluginKey as tablePluginKey,
   getPluginState,