+++ packages/editor/editor-core/src/plugins/table/index.tsx
@@ -2,13 +2,8 @@ import * as React from 'react';
 import { tableEditing } from 'prosemirror-tables';
 import { createTable } from 'prosemirror-utils';
 import TableIcon from '@atlaskit/icon/glyph/editor/table';
-import {
-  table,
-  tableCell,
-  tableHeader,
-  tableRow,
-  tableCellMinWidth,
-} from '@atlaskit/editor-common';
+import { tableCellMinWidth } from '@atlaskit/editor-common';
+import { table, tableCell, tableHeader, tableRow } from '@atlaskit/adf-schema';
 
 import LayoutButton from './ui/LayoutButton';
 import { EditorPlugin } from '../../types';