+++ packages/editor/editor-jira-transformer/src/__tests__/unit/code-block.ts
@@ -4,7 +4,7 @@ import {
   checkEncode,
   checkParse,
 } from './_test-helpers';
-import { createJIRASchema } from '@atlaskit/editor-common';
+import { createJIRASchema } from '@atlaskit/adf-schema';
 
 const schema = createJIRASchema({ allowCodeBlock: true });