+++ packages/editor/renderer/src/email/index.ts
@@ -5,7 +5,7 @@ import { Fragment, Node as PMNode, Schema } from 'prosemirror-model';
 import { Serializer } from '../serializer';
 import { nodeSerializers } from './serializers';
 import { serializeStyle } from './util';
-import { calcTableColumnWidths } from '@atlaskit/editor-common';
+import { calcTableColumnWidths } from '@atlaskit/adf-schema';
 
 const serializeNode = (node: PMNode, serializedHTML?: string): string => {
   // ignore nodes with unknown type