+++ packages/search/global-search/src/i18n/pt_BR.ts
@@ -16,7 +16,7 @@ export default {
     'Ou use a <a href={url}>Pesquisa avançada</a> (Shift + Enter) para restringir seus resultados.',
   'global_search.no_results_title': 'Nenhum resultado em títulos e pessoas',
   'global_search.no_results_body':
-    'Tente pesquisar usando outro termo. Você também pode pesquisar conteúdo da página com a pesquisa avançada ou pesquisar pessoas abaixo.',
+    'Tente novamente usando outro termo ou refine seus resultados com nossa pesquisa avançada.',
   'global_search.people.recent_people_heading': 'Recentemente trabalhou com',
   'global_search.people.people_heading': 'Pessoas',
   'global_search.people.advanced_search': 'Procurar pessoas',
@@ -28,12 +28,12 @@ export default {
   'global_search.jira.recent_issues_heading': 'Itens recentes',
   'global_search.jira.recent_people_heading': 'Trabalhou recentemente com',
   'global_search.jira.recent_containers':
-    'Painéis, filtros e projetos recentes',
+    'Painéis, projetos e filtros recentes',
   'global_search.jira.search_result_issues_heading': 'Itens',
   'global_search.jira.search_result_containers_heading':
-    'Painéis, filtros e projetos',
+    'Painéis, projetos e filtros',
   'global_search.jira.search_result_people_heading': 'Pessoas',
-  'global_search.jira.advanced_search': 'Ir para a pesquisa avançada de',
+  'global_search.jira.advanced_search': 'Pesquisa avançada para',
   'global_search.jira.advanced_search_issues': 'itens',
   'global_search.jira.advanced_search_projects': 'projetos',
   'global_search.jira.advanced_search_boards': 'painéis',
@@ -41,9 +41,16 @@ export default {
   'global_search.jira.advanced_search_people': 'pessoas',
   'global_search.jira.advanced_issue_search': 'Pesquisa avançada de itens',
   'global_search.jira.no_results_title':
-    'Nenhum resultado em resumos de itens, nomes e pessoas',
+    'Não encontramos nenhum resultado para sua pesquisa',
   'global_search.jira.no_results_body':
-    'Tente pesquisar usando outro termo. Você também pode pesquisar todos os campos de texto em itens com a pesquisa avançada ou pesquisar outros elementos de modo mais específico.',
+    'Tente novamente usando outro termo ou refine seus resultados com nossa pesquisa avançada.',
   'global_search.jira.no_recent_activity_body':
     'Ou refine seus resultados com nossa pesquisa avançada.',
+  'global_search.jira.project_type.business_project': 'Projeto de negócios',
+  'global_search.jira.project_type.software_project': 'Projeto de software',
+  'global_search.jira.project_type.service_desk_project':
+    'Projeto de central de atendimento',
+  'global_search.jira.project_type.ops_project': 'Projeto de operações',
+  'global_search.jira.result_type.board': 'Painel',
+  'global_search.jira.result_type.filter': 'Filtro',
 };