+++ packages/search/global-search/src/i18n/sk.ts
@@ -16,7 +16,7 @@ export default {
     'Alebo použite <a href={url}>rozšírené vyhľadávanie</a> (shift + enter`), aby ste upresnili výsledky.',
   'global_search.no_results_title': 'Žiadne výsledky v názvoch alebo ľuďoch',
   'global_search.no_results_body':
-    'Skúste vyhľadať iný pojem. Vďaka funkcii rozšíreného vyhľadávania môžete prehľadať obsah stránky alebo vyhľadávať ľudí.',
+    'Skúste to znova s iným pojmom alebo upresnite výsledky pomocou rozšíreného vyhľadávania.',
   'global_search.people.recent_people_heading': 'Posledná spolupráca',
   'global_search.people.people_heading': 'Ľudia',
   'global_search.people.advanced_search': 'Vyhľadať ľudí',
@@ -28,12 +28,12 @@ export default {
   'global_search.jira.recent_issues_heading': 'Nedávne problémy',
   'global_search.jira.recent_people_heading': 'Posledná spolupráca',
   'global_search.jira.recent_containers':
-    'Najnovšie nástenky, filtre a projekty',
+    'Najnovšie nástenky, projekty a filtre',
   'global_search.jira.search_result_issues_heading': 'Problémy',
   'global_search.jira.search_result_containers_heading':
-    'Nástenky, filtre a projekty',
+    'Nástenky, projekty a filtre',
   'global_search.jira.search_result_people_heading': 'Ľudia',
-  'global_search.jira.advanced_search': 'Rozšírené vyhľadávanie',
+  'global_search.jira.advanced_search': 'Rozšírené vyhľadávanie pre',
   'global_search.jira.advanced_search_issues': 'problémy',
   'global_search.jira.advanced_search_projects': 'projekty',
   'global_search.jira.advanced_search_boards': 'nástenky',
@@ -42,9 +42,16 @@ export default {
   'global_search.jira.advanced_issue_search':
     'Rozšírené vyhľadávanie problémov',
   'global_search.jira.no_results_title':
-    'Žiadne výsledky v súhrnoch problémov, názvoch a ľuďoch',
+    'Nenašli sme žiadne výsledky, ktoré by zodpovedali vášmu vyhľadávaniu',
   'global_search.jira.no_results_body':
-    'Skúste vyhľadať iný pojem. Vďaka funkcii rozšíreného vyhľadávania môžete prehľadať všetky textové polia alebo konkrétnejšie vyhľadať iné položky.',
+    'Skúste to znova s iným pojmom alebo upresnite výsledky pomocou rozšíreného vyhľadávania.',
   'global_search.jira.no_recent_activity_body':
     'Alebo svoje výsledky upresnite pomocou funkcie rozšíreného vyhľadávania.',
+  'global_search.jira.project_type.business_project': 'Obchodný projekt',
+  'global_search.jira.project_type.software_project': 'Softvérový projekt',
+  'global_search.jira.project_type.service_desk_project':
+    'Projekt technickej podpory',
+  'global_search.jira.project_type.ops_project': 'Prevádzkový projekt',
+  'global_search.jira.result_type.board': 'Nástenka',
+  'global_search.jira.result_type.filter': 'Filter',
 };