+++ packages/editor/adf-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/adf-utils
 
+## 5.4.0
+- [minor] [1205725](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1205725):
+
+  - Move schema to its own package
+
+## 5.3.4
+- Updated dependencies [80f765b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/80f765b):
+  - @atlaskit/editor-common@23.0.0
+
 ## 5.3.3
 - [patch] [34df084](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/34df084):